introduce string unification script
Python provides the user with the choice of two different quotation
signs for strings: single and double quotes. There is no practical
difference and it purely boils down to a matter of taste what to choose.
However, often enough the different syntaxes are mixed, making code look
inconsistent.
This change introduces a script that detects and possibly fixes strings
using single quotes in an attempt to enforce usage of double quotation
marks throughout all Python programs.