properly handle multi-line strings
Our approach of using a regular expression for the actual quote
replacement is fine but for multi-line strings to be handled correctly
the regular expression should not match on a per-line basis.
This change addresses the issue. Furthermore, we can no longer cache the
content of the file on a per-line basis because that does not work with
our replacement of the fixed-up strings. Instead, with this patch we
store the content as it is read as a bytes object.