change string replacement approach
Although our algorithm for replacing/unifying quotation marks works in
most cases, it turns out that there is one inherent deficiency that it
cannot cope with: Because we effectively use a search-and-replace
approach for reflecting replacements made to a string we got supplied by
the tokenizer in the buffer of data we read from the file, we can get
false replacements or miss others in case a (logical) line contains
multiple strings that are (true) prefixes of each other.
To fix this problem, this patch changes the approach for replacing once
again: instead of doing a search of the string that got replaced (or
that is to be replaced) to one that uses the information supplied from
the tokenizer to directly deduce which area in the buffer to work on.
This way, we avoid the prefix problem decribed before that is inherent
to search-and-replace.