rework internal logic to work with strings
With certain Python files we experienced assertion failures during the
unification process. Root causing them has led to the insight that
although tokenization happens on a bytes array, the reported tokens are
strings. As such, len() invocations on them would potentially yield
different results for unicode strings outside of the ASCII range than
for bytes objects. Unfortunately, it is exactly bytes objects that we
store in the internal line buffer that we use.
To fix this problem this patch causes the unifier now work with strings
internally.